[UPD] Fitness now is with double values. Fitness sharing implemented (niching technique). RAM for Boxing full analyzed with custom method implemented in ALE that allows to set the values on RAM (it is implemented locally! just for testing purposes, it will not work if it's not implemented).

+++ src/Genetic/DNA.cpp
@@ -49,7 +49,7 @@ void DNA::setCrossoverRate(double crossoverRate)
  *  that is going to get the crossover, not
  *  a.mutation
  */
-DNA DNA::crossover(const DNA &a, int ownFitness, int aFitness)
+DNA DNA::crossover(const DNA &a, double ownFitness, double aFitness)
 {
     //DNA res(a.genes, this->mutationRate);
     DNA res(a.genes.rows(), a.genes.cols());
